Resolve build issue from firmware top-level directory
When make is called from the modules directory, all is well—except that
MODULAR=n has no effect there. When building from the top-level,
MODULAR=n works but there is some other glitch which is not encountered
usually. This *seems* to resolve the glitch, but since the faulty code
went to release for 0.6.0, there must be some other more-correct
solution.